Move logs implementations to package
